Move PendingChain to Illuminate\Bus so it can be used with Lumen
